:truck: Rename common dir to shared
