+++ packages/cli/test/shared/augmentation.d.ts
